% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na.R
\name{check_na}
\alias{check_na}
\title{Check NA values
The function checks if the tibble contains NA values. Give as an output a tibble with variables that have NA values at the amount of this.}
\usage{
check_na(df)
}
\arguments{
\item{df}{A tibble or dataframe}
}
\value{
A tibble with the variables that contain NA values.
}
\description{
Check NA values
The function checks if the tibble contains NA values. Give as an output a tibble with variables that have NA values at the amount of this.
}
