% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_calc.R
\name{et_calc}
\alias{et_calc}
\title{ET calc}
\usage{
et_calc(skip = 9, volume = 2.197, area = 1.69)
}
\arguments{
\item{skip}{integer: the number of lines of the data file to skip
before beginning to read data.}

\item{volume}{The volume of the chamber in cubic meters.}

\item{area}{The area of the chamber in square meters.}
}
\value{
It will produce a plot of the data, queery the user if
they would like to modify the time interval over which
the data is fit, then print the paste statement of data
values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is meant to perform a linear and non-linear fitting
of \verb{H2O absorpotion}, using data from LiCOR machines to provide
estimates of evapotranspiraiton. It is modeled off of a previous
function designed to perform similar fitting of \code{CO2} data to
estimate \verb{net ecosystem exchange}.

Transpiration is only plants when stomata are open (daytime
measurements). Evapotranspiration is everything, plants and
soils (day time measurements), and evaporation is only soils,
(night time or dark/respiration measurements).  Thus, we are
measuring, directly, only evapotranspiration \code{(ET)} and
evaporation \code{(E)}. But, we can extract transpiration \code{(T)} as:

\code{ET = E + T}

Some report \code{T/ET} ratios. Our estimates appear reasonable with
others reported by Wang et al. and Burba et al.  Do see
Wang et al.'s review paper for comments on reliability of
nightime measurements as proxy for evaporation.  Specifically,
he quotes studies that found measurable non-zero transpiration
during plant respiration.
}
\references{
Wang, L., Niu, S., Good , S.P.,Soderberg K., McCabe, M.F.,
Sherry, R.A., Luo, Y., Zhou, X., Xia, J., Caylor, K.K. (2013). The effect of warming on grassland
evapotranspiration partitioning using laser-based isotope monitoring
techniques. https://doi.org/10.1016/j.gca.2012.12.047.

Burba, G., Schmidt, A., Scott, R.L., Nakai, T., Kathilankal, J.,
Fratini, G., Hanson, C., Law, B., McDermitt, D.K., Eckles, R.,
Furtaw, M. and Velgersdyk, M. (2012), Calculating CO2 and H2O eddy
covariance fluxes from an enclosed gas analyzer using an instantaneous
mixing ratio. Glob. Change Biol., 18: 385-399. https://doi.org/10.1111/j.1365-2486.2011.02536.x

Saleska, S.R., Harte, J. and Torn, M.S. (1999), The effect of
experimental ecosystem warming on CO2 fluxes in a montane meadow.
Global Change Biology, 5: 125-141. https://doi.org/10.1046/j.1365-2486.1999.00216.x
}
