% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nee_et_calc.R
\name{nee_et_calc}
\alias{nee_et_calc}
\title{NEE an ET calc}
\usage{
nee_et_calc(path = NULL, param = "et", skip = 9, vol = 2.197, area = 1.69)
}
\arguments{
\item{path}{files path name.}

\item{param}{chose fitting model "nee" or "et".}

\item{skip}{integer: the number of lines of the data file to skip
before beginning to read data.}

\item{vol}{The volume of the chamber in cubic meters.}

\item{area}{The area of the chamber in square meters.}
}
\value{
It will produce a plot of the data, queery the user if
they would like to modify the time interval over which
the data is fit, then print the paste statement of data
values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is meant to perform the linear and non-linear fitting to
\code{NEE} data.
}
\author{
Alex Brummer
Identify all file names corresponding to photosynthesis and respiration measurements
For reading the .txt files, replace read.csv with read.table, and add a skip = 9 parameter to the function.
}
