% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{Read Licor files}
\usage{
read_files(path,
           photo = "photo",
           resp = "resp",
           ambient = "a")
}
\arguments{
\item{path}{a path to a file.}

\item{photo}{a pattern to identify photosynthesis measurement files.}

\item{resp}{a pattern to identify respiration measurement files.}

\item{ambient}{a pattern to identify ambient flux measurement files.}
}
\value{
a list.
}
\description{
Generate a list of Licor files used in the NEE or ET measuring.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
