% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tent_data.R
\name{read_tent_data}
\alias{read_tent_data}
\title{read_tent_data}
\usage{
read_tent_data(year, ratio)
}
\arguments{
\item{year}{path to year folder where all sites folders are hosted}

\item{ratio}{Argument with two possible values "NEE" and "ET"}
}
\value{
a tibble
}
\description{
read_tent_data
}
\examples{
# Basic usage
read_tent_data(year = "reformated_files/2003", ratio = "ET")
}
