% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ints.R
\name{str_delete_before}
\alias{str_delete_before}
\title{Delete characters in a string which occur before a specified pattern}
\usage{
str_delete_before(string, pattern, pos = 1)
}
\arguments{
\item{string}{The string from which to extract.}

\item{pattern}{Pattern to look for. The default interpretation is a regular expression.}

\item{pos}{Position of the character used as a marker.}
}
\value{
The part of the string after the pattern.
}
\description{
Vectorised over 'string' and 'pattern'. from PaulESantos/pesa.
}
