% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et_calc.R
\name{et_calc}
\alias{et_calc}
\title{ET calc}
\usage{
et_calc(skip = 9, vol = 2.197, area = 1.69)
}
\arguments{
\item{skip}{integer: the number of lines of the data file to skip
before beginning to read data}

\item{vol}{The volume of the chamber in cubic meters}

\item{area}{The area of the chamber in square meters}
}
\value{
It will produce a plot of the data, queery the user if
they would like to modify the time interval over which
the data is fit, then print the paste statement of data
values
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is meant to perform a linear and non-linear fitting
of \verb{H2O absorption}, using data from LiCOR machines to provide
estimates of evapotranspiration
}
\author{
Alex Brummer
}
