% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_files}
\alias{read_files}
\title{Read LiCOR Files}
\usage{
read_files(path, photo = "photo", resp = "resp", ambient = "a")
}
\arguments{
\item{path}{A character string representing the path to the folder containing
the LiCOR files to be analyzed.}

\item{photo}{A character string representing the pattern to identify
photosynthesis measurement files.}

\item{resp}{A character string representing the pattern to identify respiration
measurement files.}

\item{ambient}{A character string representing the pattern to identify ambient
flux measurement files.}
}
\value{
A list containing three elements: 'photo_names' for photosynthesis
files, 'ambient_names' for ambient flux files, and 'resp_names' for respiration
files.
}
\description{
Generate a list of LiCOR files used in the NEE (Net Ecosystem Exchange) or
ET (Evapotranspiration) measurements.
}
\details{
This function reads LiCOR files generated by a LiCOR 7500 analyzer and organizes
them into lists based on the specified patterns.

This function reads LiCOR files based on the specified patterns for photosynthesis,
respiration, and ambient flux measurements.

If the 'path' parameter is missing, the function will prompt the user to set the
working directory to the folder containing the LiCOR files.
}
\examples{
# Example 1: Read LiCOR files from the current working directory interactively.
# li_cor_data <- read_files()

# Example 2: Read LiCOR files from a specified directory with custom file patterns.
# li_cor_data <- read_files("path/to/files",
# photo = "photosynth",
# resp = "resp_flux",
# ambient = "amb")

}
